Converted clients[] array into pointers to PKG connections, rather than container structs with a duplicate of the fd.
Fixes the problem that Keith reported, where fbserv as started from
inetd.conf would connect with client, but not take any data.
In that case, fd==0, which was used as the marker for an empty slot
in the clients[] array.  Thus, the select() input indication would
always be skipped over, and no PKGs would be read off the network.


git-svn-id: https://svn.code.sf.net/p/brlcad/code/brlcad/trunk@19037 2f96ce8b-6d43-0410-b8df-bffccc660ffb