run gqa in addition to rtcheck.
this is so that users aren't surprised by gqa reporting additional
overlaps when run as a secondary validation test.  gqa output is
treated equivalently (as if it were across 16 views instead of 3) to
help keep it's overweighted counts in check.  this reduces gqa volume
estimates, but lets them contribute to the total so they will be
listed.

also turning off per-pair plotting for now until there's a use, but
increasing default rtcheck grid size to 1024 since it was missing
a lot of things gqa was finding.



git-svn-id: https://svn.code.sf.net/p/brlcad/code/brlcad/trunk@69531 2f96ce8b-6d43-0410-b8df-bffccc660ffb