Support use of std::isnan in VS 2012.
Visual Studio 2012 is currently supported for Creo 3, but math.h's isnan is
one of the C99 functions not implemented before VS 2013, so it must be
replaced with _isnan from float.h. In order to reference std::isnan, we must
take the additional steps in the Windows config of including float.h and
defining a std::isnan wrapping _isnan.


git-svn-id: https://svn.code.sf.net/p/brlcad/code/brlcad/trunk@70494 2f96ce8b-6d43-0410-b8df-bffccc660ffb