Add toggle to draw only first unioned solid.
Add checkbutton (set the same as -F option on startup) to toggle drawing only
the first union in overlap paths.

Going with a checkbutton in a labelframe for the moment as a compromise between
many considerations and future planned work. A toggle for showing overlap
volume plots can be added as another checkbutton inside the labelframe.


git-svn-id: https://svn.code.sf.net/p/brlcad/code/brlcad/trunk@70823 2f96ce8b-6d43-0410-b8df-bffccc660ffb