Add WIP code for identifying hanging nodes.
The goal is to find the subset of unmatched edges in a triangle mesh which are closed -- not by a single matching edge from an adjacent face -- but by a chain of other edges with the same endpoints whose internal vertices are collinear (within tolerance) of the longer unmatched edge.

The internal vertices of the chain are "hanging nodes", also known as T-junctions, as they aren't topologically connected to the rest of the mesh. They represent topological holes that still leave the actual geometry watertight.

git-svn-id: https://svn.code.sf.net/p/brlcad/code/brlcad/trunk@71176 2f96ce8b-6d43-0410-b8df-bffccc660ffb