significantly rework the ogl framebuffer logic.  rewrite how shared memory segments are handled taking care to ensure the title accurately reflects what is being used, but also now limiting shared memory use to the default 512x512 image sizes (and smaller).  this is for a few reasons, one being we cannot easily determine the maximum amount of available shared memory (at least on non-linux), another being stack corruption when the limit is exceeded despite functions returning successfully (arguably a platform bug, but undetectable).  this fixes the corruption.  added some additional debugging too.
backgraound info: mac os x defaults to 4M, which with the trailing
color map would be a max of 1023x1023, so we can't even bump up to the
next power of two without handling colormappings differently (e.g., in
a different segment).

that all said, we should probably eliminate this usage of shared
memory.  it's technically bad practice from a security and data
integrity standpoint given how it's used here in that one user gets to
see another user's framebuffer if they have accounts on the same
system, and can even modify/corrupt each other's image.



git-svn-id: https://svn.code.sf.net/p/brlcad/code/brlcad/trunk@71753 2f96ce8b-6d43-0410-b8df-bffccc660ffb