Define (but disable) the mapped file parallel free test.
This test is intended to simulate what can happen with lots of db_open/db_close
calls in parallel, since db_close calls the mappedfile free call.

I believe the newest librt cache logic avoids this issue for now, and there
dont' seem to be any other calls to bu_free_mapped_files other than db_close
and the test logic, so I think we can postpone addressing this for now.

That derth of bu_free_mapped_files calls is another issue - if the expectation
is that mapped_file callers clean up after a close then db_close appears to be
the only one doing it - but barring massive memory or file handle consumption
it's not a problem that will impact users.  There are a few places where it is
probably safe to add the call now (when it can be done after parallel
processing is complete) but I'm going to wait even on those until after the
patch release to avoid any last minute unintended breakages.



git-svn-id: https://svn.code.sf.net/p/brlcad/code/brlcad/trunk@72810 2f96ce8b-6d43-0410-b8df-bffccc660ffb